+++ bmc-common/src/main/java/com/oracle/bmc/auth/internal/GetResourcePrincipalSessionTokenRequest.java
@@ -1,0 +1,24 @@
+/**
+ * Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
+ * This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
+ */
+package com.oracle.bmc.auth.internal;
+
+import javax.annotation.concurrent.Immutable;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+
+@EqualsAndHashCode(callSuper = false)
+@Immutable
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@AllArgsConstructor
+@Getter
+public class GetResourcePrincipalSessionTokenRequest {
+    private final String resourcePrincipalToken;
+    private final String servicePrincipalSessionToken;
+    private final String sessionPublicKey;
+}