+++ bmc-dataflow/src/main/java/com/oracle/bmc/dataflow/DataFlow.java
@@ -97,7 +97,7 @@ public interface DataFlow extends AutoCloseable {
     CreateApplicationResponse createApplication(CreateApplicationRequest request);
 
     /**
-     * Creates a private endpoint to be used by an application.
+     * Creates a private endpoint to be used by applications.
      *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
@@ -208,7 +208,7 @@ public interface DataFlow extends AutoCloseable {
     GetWorkRequestResponse getWorkRequest(GetWorkRequestRequest request);
 
     /**
-     * Lists all applications in the specified compartment.
+     * Lists all applications in the specified compartment. Only one parameter other than compartmentId may also be included in a query. The query must include compartmentId. If the query does not include compartmentId, or includes compartmentId but two or more other parameters an error is returned.
      *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
@@ -241,7 +241,7 @@ public interface DataFlow extends AutoCloseable {
     ListRunLogsResponse listRunLogs(ListRunLogsRequest request);
 
     /**
-     * Lists all runs of an application in the specified compartment.
+     * Lists all runs of an application in the specified compartment.  Only one parameter other than compartmentId may also be included in a query. The query must include compartmentId. If the query does not include compartmentId, or includes compartmentId but two or more other parameters an error is returned.
      *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation