+++ bmc-dataflow/src/main/java/com/oracle/bmc/dataflow/model/CreateRunDetails.java
@@ -7,17 +7,33 @@ package com.oracle.bmc.dataflow.model;
 /**
  * The create run details. The following properties are optional and override the default values
  * set in the associated application:
+ *   - applicationId
+ *   - archiveUri
  *   - arguments
  *   - configuration
  *   - definedTags
+ *   - displayName
  *   - driverShape
+ *   - execute
  *   - executorShape
  *   - freeformTags
  *   - logsBucketUri
  *   - numExecutors
  *   - parameters
+ *   - sparkVersion
  *   - warehouseBucketUri
- * If the optional properties are not specified, they are copied over from the parent application.
+ * It is expected that either the applicationId or the execute parameter is specified; but not both.
+ * If both or none are set, a Bad Request (HTTP 400) status will be sent as the response.
+ * If an appicationId is not specified, then a value for the execute parameter is expected.
+ * Using data parsed from the value, a new application will be created and assicated with the new run.
+ * See information on the execute parameter for details on the format of this parameter.
+ * <p>
+ * The optional parameter spark version can only be specified when using the execute parameter.  If it
+ * is not specified when using the execute parameter, the latest version will be used as default.
+ * If the execute parameter is not used, the spark version will be taken from the associated application.
+ * <p>
+ * If displayName is not specified, it will be derived from the displayName of associated application or
+ * set by API using fileUri's application file name.
  * Once a run is created, its properties (except for definedTags and freeformTags) cannot be changed.
  * If the parent application's properties (including definedTags and freeformTags) are updated,
  * the corresponding properties of the run will not update.
@@ -46,6 +62,15 @@ public class CreateRunDetails {
         public Builder applicationId(String applicationId) {
             this.applicationId = applicationId;
             this.__explicitlySet__.add("applicationId");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonProperty("archiveUri")
+        private String archiveUri;
+
+        public Builder archiveUri(String archiveUri) {
+            this.archiveUri = archiveUri;
+            this.__explicitlySet__.add("archiveUri");
             return this;
         }
 
@@ -104,6 +129,15 @@ public class CreateRunDetails {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("execute")
+        private String execute;
+
+        public Builder execute(String execute) {
+            this.execute = execute;
+            this.__explicitlySet__.add("execute");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("executorShape")
         private String executorShape;
 
@@ -149,6 +183,15 @@ public class CreateRunDetails {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("sparkVersion")
+        private String sparkVersion;
+
+        public Builder sparkVersion(String sparkVersion) {
+            this.sparkVersion = sparkVersion;
+            this.__explicitlySet__.add("sparkVersion");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("warehouseBucketUri")
         private String warehouseBucketUri;
 
@@ -165,17 +208,20 @@ public class CreateRunDetails {
             CreateRunDetails __instance__ =
                     new CreateRunDetails(
                             applicationId,
+                            archiveUri,
                             arguments,
                             compartmentId,
                             configuration,
                             definedTags,
                             displayName,
                             driverShape,
+                            execute,
                             executorShape,
                             freeformTags,
                             logsBucketUri,
                             numExecutors,
                             parameters,
+                            sparkVersion,
                             warehouseBucketUri);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
             return __instance__;
@@ -185,17 +231,20 @@ public class CreateRunDetails {
         public Builder copy(CreateRunDetails o) {
             Builder copiedBuilder =
                     applicationId(o.getApplicationId())
+                            .archiveUri(o.getArchiveUri())
                             .arguments(o.getArguments())
                             .compartmentId(o.getCompartmentId())
                             .configuration(o.getConfiguration())
                             .definedTags(o.getDefinedTags())
                             .displayName(o.getDisplayName())
                             .driverShape(o.getDriverShape())
+                            .execute(o.getExecute())
                             .executorShape(o.getExecutorShape())
                             .freeformTags(o.getFreeformTags())
                             .logsBucketUri(o.getLogsBucketUri())
                             .numExecutors(o.getNumExecutors())
                             .parameters(o.getParameters())
+                            .sparkVersion(o.getSparkVersion())
                             .warehouseBucketUri(o.getWarehouseBucketUri());
 
             copiedBuilder.__explicitlySet__.retainAll(o.__explicitlySet__);
@@ -211,11 +260,19 @@ public class CreateRunDetails {
     }
 
     /**
-     * The application ID.
+     * The OCID of the associated application. If this value is set, then no value for the execute parameter is required. If this value is not set, then a value for the execute parameter is required, and a new application is created and associated with the new run.
      *
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("applicationId")
     String applicationId;
+
+    /**
+     * An Oracle Cloud Infrastructure URI of an archive.zip file containing custom dependencies that may be used to support the execution a Python, Java, or Scala application.
+     * See https://docs.cloud.oracle.com/iaas/Content/API/SDKDocs/hdfsconnector.htm#uriformat.
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("archiveUri")
+    String archiveUri;
 
     /**
      * The arguments passed to the running application as command line arguments.  An argument is
@@ -258,7 +315,7 @@ public class CreateRunDetails {
     java.util.Map<String, java.util.Map<String, Object>> definedTags;
 
     /**
-     * A user-friendly name. It does not have to be unique. Avoid entering confidential information.
+     * A user-friendly name that does not have to be unique. Avoid entering confidential information. If this value is not specified, it will be derived from the associated application's displayName or set by API using fileUri's application file name.
      *
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("displayName")
@@ -270,6 +327,17 @@ public class CreateRunDetails {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("driverShape")
     String driverShape;
+
+    /**
+     * The input used for spark-submit command. For more details see https://spark.apache.org/docs/latest/submitting-applications.html#launching-applications-with-spark-submit.
+     * Supported options include ``--class``, ``--file``, ``--jars``, ``--conf``, ``--py-files``, and main application file with arguments.
+     * Example: ``--jars oci://path/to/a.jar,oci://path/to/b.jar --files oci://path/to/a.json,oci://path/to/b.csv --py-files oci://path/to/a.py,oci://path/to/b.py --conf spark.sql.crossJoin.enabled=true --class org.apache.spark.examples.SparkPi oci://path/to/main.jar 10``
+     * Note: If execute is specified together with applicationId, className, configuration, fileUri, language, arguments, parameters during application create/update, or run create/submit,
+     * Data Flow service will use derived information from execute input only.
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("execute")
+    String execute;
 
     /**
      * The VM shape for the executors. Sets the executor cores and memory.
@@ -311,6 +379,13 @@ public class CreateRunDetails {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("parameters")
     java.util.List<ApplicationParameter> parameters;
+
+    /**
+     * The Spark version utilized to run the application. This value may be set if applicationId is not since the Spark version will be taken from the associated application.
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("sparkVersion")
+    String sparkVersion;
 
     /**
      * An Oracle Cloud Infrastructure URI of the bucket to be used as default warehouse directory