+++ bmc-resourcemanager/src/main/java/com/oracle/bmc/resourcemanager/model/ConfigSource.java
@@ -34,6 +34,10 @@ package com.oracle.bmc.resourcemanager.model;
         name = "GIT_CONFIG_SOURCE"
     ),
     @com.fasterxml.jackson.annotation.JsonSubTypes.Type(
+        value = ObjectStorageConfigSource.class,
+        name = "OBJECT_STORAGE_CONFIG_SOURCE"
+    ),
+    @com.fasterxml.jackson.annotation.JsonSubTypes.Type(
         value = CompartmentConfigSource.class,
         name = "COMPARTMENT_CONFIG_SOURCE"
     ),
@@ -63,6 +67,7 @@ public class ConfigSource {
         ZipUpload("ZIP_UPLOAD"),
         GitConfigSource("GIT_CONFIG_SOURCE"),
         CompartmentConfigSource("COMPARTMENT_CONFIG_SOURCE"),
+        ObjectStorageConfigSource("OBJECT_STORAGE_CONFIG_SOURCE"),
 
         /**
          * This value is used if a service returns a value for this enum that is not recognized by this