+++ bmc-resourcemanager/src/main/java/com/oracle/bmc/resourcemanager/model/ObjectStorageConfigSource.java
@@ -1,0 +1,133 @@
+/**
+ * Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
+ * This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
+ */
+package com.oracle.bmc.resourcemanager.model;
+
+/**
+ * Metadata about the Object Storage configuration source.
+ *
+ * <br/>
+ * Note: Objects should always be created or deserialized using the {@link Builder}. This model distinguishes fields
+ * that are {@code null} because they are unset from fields that are explicitly set to {@code null}. This is done in
+ * the setter methods of the {@link Builder}, which maintain a set of all explicitly set fields called
+ * {@link #__explicitlySet__}. The {@link #hashCode()} and {@link #equals(Object)} methods are implemented to take
+ * {@link #__explicitlySet__} into account. The constructor, on the other hand, does not set {@link #__explicitlySet__}
+ * (since the constructor cannot distinguish explicit {@code null} from unset {@code null}).
+ **/
+@javax.annotation.Generated(value = "OracleSDKGenerator", comments = "API Version: 20180917")
+@lombok.Value
+@com.fasterxml.jackson.databind.annotation.JsonDeserialize(
+    builder = ObjectStorageConfigSource.Builder.class
+)
+@lombok.ToString(callSuper = true)
+@lombok.EqualsAndHashCode(callSuper = true)
+@com.fasterxml.jackson.annotation.JsonTypeInfo(
+    use = com.fasterxml.jackson.annotation.JsonTypeInfo.Id.NAME,
+    include = com.fasterxml.jackson.annotation.JsonTypeInfo.As.PROPERTY,
+    property = "configSourceType"
+)
+@com.fasterxml.jackson.annotation.JsonFilter(com.oracle.bmc.http.internal.ExplicitlySetFilter.NAME)
+@lombok.Builder(builderClassName = "Builder", toBuilder = true)
+public class ObjectStorageConfigSource extends ConfigSource {
+    @com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder(withPrefix = "")
+    @lombok.experimental.Accessors(fluent = true)
+    public static class Builder {
+        @com.fasterxml.jackson.annotation.JsonProperty("workingDirectory")
+        private String workingDirectory;
+
+        public Builder workingDirectory(String workingDirectory) {
+            this.workingDirectory = workingDirectory;
+            this.__explicitlySet__.add("workingDirectory");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonProperty("region")
+        private String region;
+
+        public Builder region(String region) {
+            this.region = region;
+            this.__explicitlySet__.add("region");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonProperty("namespace")
+        private String namespace;
+
+        public Builder namespace(String namespace) {
+            this.namespace = namespace;
+            this.__explicitlySet__.add("namespace");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonProperty("bucketName")
+        private String bucketName;
+
+        public Builder bucketName(String bucketName) {
+            this.bucketName = bucketName;
+            this.__explicitlySet__.add("bucketName");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonIgnore
+        private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
+
+        public ObjectStorageConfigSource build() {
+            ObjectStorageConfigSource __instance__ =
+                    new ObjectStorageConfigSource(workingDirectory, region, namespace, bucketName);
+            __instance__.__explicitlySet__.addAll(__explicitlySet__);
+            return __instance__;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonIgnore
+        public Builder copy(ObjectStorageConfigSource o) {
+            Builder copiedBuilder =
+                    workingDirectory(o.getWorkingDirectory())
+                            .region(o.getRegion())
+                            .namespace(o.getNamespace())
+                            .bucketName(o.getBucketName());
+
+            copiedBuilder.__explicitlySet__.retainAll(o.__explicitlySet__);
+            return copiedBuilder;
+        }
+    }
+
+    /**
+     * Create a new builder.
+     */
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    @Deprecated
+    public ObjectStorageConfigSource(
+            String workingDirectory, String region, String namespace, String bucketName) {
+        super(workingDirectory);
+        this.region = region;
+        this.namespace = namespace;
+        this.bucketName = bucketName;
+    }
+
+    /**
+     * The name of the bucket's region.
+     * Example: `PHX`
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("region")
+    String region;
+
+    /**
+     * The Object Storage namespace that contains the bucket.
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("namespace")
+    String namespace;
+
+    /**
+     * The name of the bucket that contains the Terraform configuration files.
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("bucketName")
+    String bucketName;
+
+    @com.fasterxml.jackson.annotation.JsonIgnore
+    private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
+}