+++ bmc-resourcemanager/src/main/java/com/oracle/bmc/resourcemanager/model/UpdateConfigSourceDetails.java
@@ -34,6 +34,10 @@ package com.oracle.bmc.resourcemanager.model;
         name = "GIT_CONFIG_SOURCE"
     ),
     @com.fasterxml.jackson.annotation.JsonSubTypes.Type(
+        value = UpdateObjectStorageConfigSourceDetails.class,
+        name = "OBJECT_STORAGE_CONFIG_SOURCE"
+    ),
+    @com.fasterxml.jackson.annotation.JsonSubTypes.Type(
         value = UpdateZipUploadConfigSourceDetails.class,
         name = "ZIP_UPLOAD"
     )